+++ MCC/reprojectionError.cpp
@@ -3,13 +3,13 @@
 
 #include "reprojectionError.h"
 
-double computeReprojectionErrors(const std::vector<std::vector<cv::Point3f> >& objectPoints, 
-	const std::vector<std::vector<cv::Point2f> >& imagePoints, 
-	const std::vector<cv::Mat>& rvecs, 
-	const std::vector<cv::Mat>& tvecs, 
-	const cv::Mat& cameraMatrix, 
-	const cv::Mat& distCoeffs, 
-	std::vector<float>& perViewErrors, 
+double computeReprojectionErrors(const std::vector<std::vector<cv::Point3f> >& objectPoints,
+	const std::vector<std::vector<cv::Point2f> >& imagePoints,
+	const std::vector<cv::Mat>& rvecs,
+	const std::vector<cv::Mat>& tvecs,
+	const cv::Mat& cameraMatrix,
+	const cv::Mat& distCoeffs,
+	std::vector<float>& perViewErrors,
 	bool fisheye)
 {
 	std::vector<cv::Point2f> imagePoints2;
@@ -40,13 +40,13 @@ double computeReprojectionErrors(const std::vector<std::vector<cv::Point3f> >& o
 }
 
 
-double computeReprojectionErrors(const std::vector<std::vector<cv::Point3f> >& objectPoints, 
-	const std::vector<std::vector<cv::Point2f> >& imagePoints, 
-	const cv::Mat& rvecs, 
-	const cv::Mat& tvecs, 
-	const cv::Mat& cameraMatrix, 
-	const cv::Mat& distCoeffs, 
-	std::vector<float>& perViewErrors, 
+double computeReprojectionErrors(const std::vector<std::vector<cv::Point3d> >& objectPoints,
+	const std::vector<std::vector<cv::Point2d> >& imagePoints,
+	const cv::Mat& rvecs,
+	const cv::Mat& tvecs,
+	const cv::Mat& cameraMatrix,
+	const cv::Mat& distCoeffs,
+	std::vector<float>& perViewErrors,
 	bool fisheye)
 {
 	std::vector<cv::Point2f> imagePoints2;
