+++ MCC/reprojectionError.h
@@ -9,8 +9,8 @@ double computeReprojectionErrors(const std::vector<std::vector<cv::Point3f> >& o
 	const cv::Mat& cameraMatrix, const cv::Mat& distCoeffs,
 	std::vector<float>& perViewErrors, bool fisheye);
 
-double computeReprojectionErrors(const std::vector<std::vector<cv::Point3f> >& objectPoints,
-	const std::vector<std::vector<cv::Point2f> >& imagePoints,
+double computeReprojectionErrors(const std::vector<std::vector<cv::Point3d> >& objectPoints,
+	const std::vector<std::vector<cv::Point2d> >& imagePoints,
 	const cv::Mat& rvecs,
 	const cv::Mat& tvecs,
 	const cv::Mat& cameraMatrix,
@@ -33,3 +33,4 @@ double computeReprojectionErrors(const std::vector<cv::Point3d>& objectPoints,
 	const cv::Mat& cameraMatrix,
 	const cv::Mat& distCoeffs,
 	bool fisheye);
+
